Simplify questionnaires list states

+++ MentalHealth/Questionnaires/QuestionnairesListView.swift
@@ -4,22 +4,30 @@ import SwiftData
 struct QuestionnairesListView: View {
     @Environment(\.modelContext) private var modelContext
     @Query private var questionnaires: [Questionnaire]
-
+    
     var body: some View {
-        List {
-            ForEach(questionnaires) { questionnaire in
-                NavigationLink {
-                    Text("Questionnaire at \(questionnaire.timestamp, format: Date.FormatStyle(date: .numeric, time: .standard))")
-                } label: {
-                    Text(questionnaire.timestamp, format: Date.FormatStyle(date: .long, time: .shortened))
+        Group {
+            if questionnaires.isEmpty {
+                QuestionnairesEmptyView()
+            } else {
+                List {
+                    ForEach(questionnaires) { questionnaire in
+                        NavigationLink {
+                            Text("Questionnaire at \(questionnaire.timestamp, format: Date.FormatStyle(date: .numeric, time: .standard))")
+                        } label: {
+                            Text(questionnaire.timestamp, format: Date.FormatStyle(date: .long, time: .shortened))
+                        }
+                    }
+                    .onDelete(perform: deleteQuestionnaires)
                 }
             }
-            .onDelete(perform: deleteQuestionnaires)
         }
         .toolbar {
 #if os(iOS)
-            ToolbarItem(placement: .navigationBarTrailing) {
-                EditButton()
+            if !questionnaires.isEmpty {
+                ToolbarItem(placement: .navigationBarTrailing) {
+                    EditButton()
+                }
             }
 #endif
             ToolbarItem {
@@ -30,14 +38,14 @@ struct QuestionnairesListView: View {
         }
         .navigationTitle("Questionnaires")
     }
-
+    
     private func addQuestionnaire() {
         withAnimation {
             let newQuestionnaire = Questionnaire(timestamp: Date())
             modelContext.insert(newQuestionnaire)
         }
     }
-
+    
     private func deleteQuestionnaires(offsets: IndexSet) {
         withAnimation {
             for index in offsets {
